csg/requirements.txt
Django==1.6.2
MySQL-python==1.2.3
PIL==1.1.7
Whoosh==2.6.0
django-haystack==2.1.0
django-markdown-deux==1.0.4
django-pagedown==0.0.5
django-simple-captcha==0.4.1
ghdiff==0.1
markdown2==2.2.0
oauthlib==0.6.1
python-openid==2.2.5
python-social-auth==0.1.22
requests==2.2.1
requests-oauthlib==0.4.0
six==1.5.2
wsgiref==0.1.2
